Split _util into extract and copyright
Kind of. This commit is a bit of a mess, but will help a lot in the
future. All stuff related to the extraction of information out of files
is now in its own module, and all stuff related to the parsing and
handling of copyright lines is also in its own module.

The problem caused by this commit is that the tests are incomplete.
Previously, not all functions in _util were tested in the understanding
that they were just small helper functions for _actual_ functions that
are under test. Now, there are public functions in public modules that
are not tested, and by chance, all of the remaining _util is currently
untested.
