+++ tests/test_extract.py
@@ -8,7 +8,7 @@
 #
 # SPDX-License-Identifier: GPL-3.0-or-later
 
-"""Tests for reuse._util"""
+"""Tests for reuse.extract"""
 
 import os
 from inspect import cleandoc
@@ -17,8 +17,13 @@ from io import BytesIO
 import pytest
 from boolean.boolean import ParseError
 
-from reuse import _util
-from reuse._util import _LICENSING
+from reuse import _LICENSING, ReuseInfo
+from reuse.extract import (
+    decoded_text_from_binary,
+    detect_line_endings,
+    extract_reuse_info,
+    filter_ignore_block,
+)
 
 # REUSE-IgnoreStart
 
@@ -27,15 +32,13 @@ def test_extract_expression():
     """Parse various expressions."""
     expressions = ["GPL-3.0+", "GPL-3.0 AND CC0-1.0", "nonsense"]
     for expression in expressions:
-        result = _util.extract_reuse_info(
-            f"SPDX-License-Identifier: {expression}"
-        )
+        result = extract_reuse_info(f"SPDX-License-Identifier: {expression}")
         assert result.spdx_expressions == {_LICENSING.parse(expression)}
 
 
 def test_extract_expression_from_ascii_art_frame():
     """Parse an expression from an ASCII art frame"""
-    result = _util.extract_reuse_info(
+    result = extract_reuse_info(
         cleandoc(
             """
              /**********************************\\
@@ -51,20 +54,20 @@ def test_extract_erroneous_expression():
     """Parse an incorrect expression."""
     expression = "SPDX-License-Identifier: GPL-3.0-or-later AND (MIT OR)"
     with pytest.raises(ParseError):
-        _util.extract_reuse_info(expression)
+        extract_reuse_info(expression)
 
 
 def test_extract_no_info():
     """Given a string without REUSE information, return an empty ReuseInfo
     object.
     """
-    result = _util.extract_reuse_info("")
-    assert result == _util.ReuseInfo()
+    result = extract_reuse_info("")
+    assert result == ReuseInfo()
 
 
 def test_extract_tab():
     """A tag followed by a tab is also valid."""
-    result = _util.extract_reuse_info("SPDX-License-Identifier:\tMIT")
+    result = extract_reuse_info("SPDX-License-Identifier:\tMIT")
     assert result.spdx_expressions == {_LICENSING.parse("MIT")}
 
 
@@ -72,14 +75,14 @@ def test_extract_many_whitespace():
     """When a tag is followed by a lot of whitespace, the whitespace should be
     filtered out.
     """
-    result = _util.extract_reuse_info("SPDX-License-Identifier:    MIT")
+    result = extract_reuse_info("SPDX-License-Identifier:    MIT")
     assert result.spdx_expressions == {_LICENSING.parse("MIT")}
 
 
 def test_extract_bibtex_comment():
     """A special case for BibTex comments."""
     expression = "@Comment{SPDX-License-Identifier: GPL-3.0-or-later}"
-    result = _util.extract_reuse_info(expression)
+    result = extract_reuse_info(expression)
     assert str(list(result.spdx_expressions)[0]) == "GPL-3.0-or-later"
 
 
@@ -88,23 +91,21 @@ def test_extract_copyright():
     information.
     """
     copyright_line = "SPDX-FileCopyrightText: 2019 Jane Doe"
-    result = _util.extract_reuse_info(copyright_line)
+    result = extract_reuse_info(copyright_line)
     assert result.copyright_lines == {copyright_line}
 
 
 def test_extract_copyright_duplicate():
     """When a copyright line is duplicated, only yield one."""
     copyright_line = "SPDX-FileCopyrightText: 2019 Jane Doe"
-    result = _util.extract_reuse_info(
-        "\n".join((copyright_line, copyright_line))
-    )
+    result = extract_reuse_info("\n".join((copyright_line, copyright_line)))
     assert result.copyright_lines == {copyright_line}
 
 
 def test_extract_copyright_tab():
     """A tag followed by a tab is also valid."""
     copyright_line = "SPDX-FileCopyrightText:\t2019 Jane Doe"
-    result = _util.extract_reuse_info(copyright_line)
+    result = extract_reuse_info(copyright_line)
     assert result.copyright_lines == {copyright_line}
 
 
@@ -113,7 +114,7 @@ def test_extract_copyright_many_whitespace():
     whitespace is not filtered out.
     """
     copyright_line = "SPDX-FileCopyrightText:    2019 Jane Doe"
-    result = _util.extract_reuse_info(copyright_line)
+    result = extract_reuse_info(copyright_line)
     assert result.copyright_lines == {copyright_line}
 
 
@@ -133,7 +134,7 @@ def test_extract_copyright_variations():
         """
     )
 
-    result = _util.extract_reuse_info(text)
+    result = extract_reuse_info(text)
     lines = text.splitlines()
     for line in lines:
         assert line in result.copyright_lines
@@ -155,7 +156,7 @@ def test_extract_with_ignore_block():
         SPDX-FileCopyrightText: 2019 Eve
         """
     )
-    result = _util.extract_reuse_info(text)
+    result = extract_reuse_info(text)
     assert len(result.copyright_lines) == 2
     assert len(result.spdx_expressions) == 1
 
@@ -165,7 +166,7 @@ def test_extract_sameline_multiline():
     do not include the comment end pattern as part of the copyright.
     """
     text = "<!-- SPDX-FileCopyrightText: Jane Doe -->"
-    result = _util.extract_reuse_info(text)
+    result = extract_reuse_info(text)
     assert len(result.copyright_lines) == 1
     assert result.copyright_lines == {"SPDX-FileCopyrightText: Jane Doe"}
 
@@ -185,7 +186,7 @@ def test_extract_special_endings():
         [Copyright 2019 Ajnulo] ::
         """
     )
-    result = _util.extract_reuse_info(text)
+    result = extract_reuse_info(text)
     for item in result.copyright_lines:
         assert ">" not in item
         assert "] ::" not in item
@@ -198,7 +199,7 @@ def test_extract_contributors():
         # SPDX-FileContributor: Jane Doe
         """
     )
-    result = _util.extract_reuse_info(text)
+    result = extract_reuse_info(text)
     assert result.contributor_lines == {"Jane Doe"}
 
 
@@ -217,7 +218,7 @@ def test_filter_ignore_block_with_comment_style():
     )
     expected = "Relevant text\n# \nOther relevant text"
 
-    result = _util.filter_ignore_block(text)
+    result = filter_ignore_block(text)
     assert result == expected
 
 
@@ -242,7 +243,7 @@ def test_filter_ignore_block_non_comment_style():
         """
     )
 
-    result = _util.filter_ignore_block(text)
+    result = filter_ignore_block(text)
     assert result == expected
 
 
@@ -267,7 +268,7 @@ def test_filter_ignore_block_with_ignored_information_on_same_line():
         """
     )
 
-    result = _util.filter_ignore_block(text)
+    result = filter_ignore_block(text)
     assert result == expected
 
 
@@ -284,7 +285,7 @@ def test_filter_ignore_block_with_relevant_information_on_same_line():
     )
     expected = "Relevant textOther relevant text"
 
-    result = _util.filter_ignore_block(text)
+    result = filter_ignore_block(text)
     assert result == expected
 
 
@@ -305,7 +306,7 @@ def test_filter_ignore_block_with_beginning_and_end_on_same_line_correct_order()
         """
     )
 
-    result = _util.filter_ignore_block(text)
+    result = filter_ignore_block(text)
     assert result == expected
 
 
@@ -316,7 +317,7 @@ def test_filter_ignore_block_with_beginning_and_end_on_same_line_wrong_order():
     text = "Relevant textREUSE-IgnoreEndOther relevant textREUSE-IgnoreStartIgnored text"  # pylint: disable=line-too-long
     expected = "Relevant textREUSE-IgnoreEndOther relevant text"
 
-    result = _util.filter_ignore_block(text)
+    result = filter_ignore_block(text)
     assert result == expected
 
 
@@ -334,7 +335,7 @@ def test_filter_ignore_block_without_end():
     )
     expected = "Relevant text\n"
 
-    result = _util.filter_ignore_block(text)
+    result = filter_ignore_block(text)
     assert result == expected
 
 
@@ -365,166 +366,49 @@ def test_filter_ignore_block_with_multiple_ignore_blocks():
         """
     )
 
-    result = _util.filter_ignore_block(text)
+    result = filter_ignore_block(text)
     assert result == expected
-
-
-def test_make_copyright_line_simple():
-    """Given a simple statement, make it a copyright line."""
-    assert _util.make_copyright_line("hello") == "SPDX-FileCopyrightText: hello"
-
-
-def test_make_copyright_line_year():
-    """Given a simple statement and a year, make it a copyright line."""
-    assert (
-        _util.make_copyright_line("hello", year="2019")
-        == "SPDX-FileCopyrightText: 2019 hello"
-    )
-
-
-def test_make_copyright_line_prefix_spdx():
-    """Given a simple statement and prefix, make it a copyright line."""
-    statement = _util.make_copyright_line("hello", copyright_prefix="spdx")
-    assert statement == "SPDX-FileCopyrightText: hello"
-
-
-def test_make_copyright_line_prefix_spdx_year():
-    """Given a simple statement, prefix and a year, make it a copyright line."""
-    statement = _util.make_copyright_line(
-        "hello", year=2019, copyright_prefix="spdx"
-    )
-    assert statement == "SPDX-FileCopyrightText: 2019 hello"
-
-
-def test_make_copyright_line_prefix_spdx_c_year():
-    """Given a simple statement, prefix and a year, make it a copyright line."""
-    statement = _util.make_copyright_line(
-        "hello", year=2019, copyright_prefix="spdx-c"
-    )
-    assert statement == "SPDX-FileCopyrightText: (C) 2019 hello"
-
-
-def test_make_copyright_line_prefix_spdx_symbol_year():
-    """Given a simple statement, prefix and a year, make it a copyright line."""
-    statement = _util.make_copyright_line(
-        "hello", year=2019, copyright_prefix="spdx-symbol"
-    )
-    assert statement == "SPDX-FileCopyrightText: © 2019 hello"
-
-
-def test_make_copyright_line_prefix_string_year():
-    """Given a simple statement, prefix and a year, make it a copyright line."""
-    statement = _util.make_copyright_line(
-        "hello", year=2019, copyright_prefix="string"
-    )
-    assert statement == "Copyright 2019 hello"
-
-
-def test_make_copyright_line_prefix_string_c_year():
-    """Given a simple statement, prefix and a year, make it a copyright line."""
-    statement = _util.make_copyright_line(
-        "hello", year=2019, copyright_prefix="string-c"
-    )
-    assert statement == "Copyright (C) 2019 hello"
-
-
-def test_make_copyright_line_prefix_spdx_string_c_year():
-    """Given a simple statement, prefix and a year, make it a copyright line."""
-    statement = _util.make_copyright_line(
-        "hello", year=2019, copyright_prefix="spdx-string-c"
-    )
-    assert statement == "SPDX-FileCopyrightText: Copyright (C) 2019 hello"
-
-
-def test_make_copyright_line_prefix_spdx_string_year():
-    """Given a simple statement, prefix and a year, make it a copyright line."""
-    statement = _util.make_copyright_line(
-        "hello", year=2019, copyright_prefix="spdx-string"
-    )
-    assert statement == "SPDX-FileCopyrightText: Copyright 2019 hello"
-
-
-def test_make_copyright_line_prefix_spdx_string_symbol_year():
-    """Given a simple statement, prefix and a year, make it a copyright line."""
-    statement = _util.make_copyright_line(
-        "hello", year=2019, copyright_prefix="spdx-string-symbol"
-    )
-    assert statement == "SPDX-FileCopyrightText: Copyright © 2019 hello"
-
-
-def test_make_copyright_line_prefix_string_symbol_year():
-    """Given a simple statement, prefix and a year, make it a copyright line."""
-    statement = _util.make_copyright_line(
-        "hello", year=2019, copyright_prefix="string-symbol"
-    )
-    assert statement == "Copyright © 2019 hello"
-
-
-def test_make_copyright_line_prefix_symbol_year():
-    """Given a simple statement, prefix and a year, make it a copyright line."""
-    statement = _util.make_copyright_line(
-        "hello", year=2019, copyright_prefix="symbol"
-    )
-    assert statement == "© 2019 hello"
-
-
-def test_make_copyright_line_existing_spdx_copyright():
-    """Given a copyright line, do nothing."""
-    value = "SPDX-FileCopyrightText: hello"
-    assert _util.make_copyright_line(value) == value
-
-
-def test_make_copyright_line_existing_other_copyright():
-    """Given a non-SPDX copyright line, do nothing."""
-    value = "© hello"
-    assert _util.make_copyright_line(value) == value
-
-
-def test_make_copyright_line_multine_error():
-    """Given a multiline argument, expect an error."""
-    with pytest.raises(RuntimeError):
-        _util.make_copyright_line("hello\nworld")
 
 
 def test_decoded_text_from_binary_simple():
     """A unicode string encoded as bytes object decodes back correctly."""
     text = "Hello, world ☺"
     encoded = text.encode("utf-8")
-    assert _util.decoded_text_from_binary(BytesIO(encoded)) == text
+    assert decoded_text_from_binary(BytesIO(encoded)) == text
 
 
 def test_decoded_text_from_binary_size():
     """Only a given amount of bytes is decoded."""
     text = "Hello, world ☺"
     encoded = text.encode("utf-8")
-    assert _util.decoded_text_from_binary(BytesIO(encoded), size=5) == "Hello"
+    assert decoded_text_from_binary(BytesIO(encoded), size=5) == "Hello"
 
 
 def test_decoded_text_from_binary_crlf():
     """Given CRLF line endings, convert to LF."""
     text = "Hello\r\nworld"
     encoded = text.encode("utf-8")
-    assert _util.decoded_text_from_binary(BytesIO(encoded)) == "Hello\nworld"
+    assert decoded_text_from_binary(BytesIO(encoded)) == "Hello\nworld"
 
 
 def test_detect_line_endings_windows():
     """Given a CRLF string, detect the line endings."""
-    assert _util.detect_line_endings("hello\r\nworld") == "\r\n"
+    assert detect_line_endings("hello\r\nworld") == "\r\n"
 
 
 def test_detect_line_endings_mac():
     """Given a CR string, detect the line endings."""
-    assert _util.detect_line_endings("hello\rworld") == "\r"
+    assert detect_line_endings("hello\rworld") == "\r"
 
 
 def test_detect_line_endings_linux():
     """Given a LF string, detect the line endings."""
-    assert _util.detect_line_endings("hello\nworld") == "\n"
+    assert detect_line_endings("hello\nworld") == "\n"
 
 
 def test_detect_line_endings_no_newlines():
     """Given a file without line endings, default to os.linesep."""
-    assert _util.detect_line_endings("hello world") == os.linesep
+    assert detect_line_endings("hello world") == os.linesep
 
 
-# REUSE-IgnoreEnd
+# Reuse-IgnoreEnd
